app/requirements.txt
openai>=1.42.0
python-dotenv>=1.0.0
